backend/requirements.txt
flask==3.1.1
pycryptodome==3.23.0
flask-cors